AppEvaluadorCliente-main/requirements.txt
flask
werkzeug
python-dotenv
google-generativeai
gspread
oauth2client
pandas
numpy
numpy-financial
flask-limiter